Move webpack to the root
