--- .dev/config/webpack.prod.js
@@ -1,12 +1,0 @@
-/* global module, require */
-
-const { merge } = require( 'webpack-merge' );
-const common = require( './webpack.common.js' );
-
-module.exports = merge( common, {
-	mode: 'production',
-
-	optimization: {
-		minimize: true,
-	},
-} );
