+++ package.json
@@ -26,10 +26,10 @@
   ],
   "scripts": {
     "setup": "composer install && yarn install && yarn build",
-    "start": "NODE_ENV=development webpack --watch --config .dev/config/webpack.dev.js",
-    "build": "NODE_ENV=production webpack --config .dev/config/webpack.prod.js",
-    "postbuild": "rm -rf temp && mkdir temp && cp -a dist/* temp && NODE_ENV=development webpack --config .dev/config/webpack.dev.js && cp -a temp/* dist && rm -rf temp && .dev/deploy-scripts/tweak-css-files.sh && yarn makepot",
-    "postdev": "mkdir temp && cp -a dist/* temp && NODE_ENV=production webpack --config .dev/config/webpack.prod.js && cp -a temp/* dist && rm -rf temp && .dev/deploy-scripts/tweak-css-files.sh",
+    "start": "NODE_ENV=development webpack --watch --config webpack.config.js",
+    "build": "NODE_ENV=production webpack --config webpack.config.js",
+    "postbuild": "rm -rf temp && mkdir temp && cp -a dist/* temp && NODE_ENV=development webpack --config webpack.config.js && cp -a temp/* dist && rm -rf temp && .dev/deploy-scripts/tweak-css-files.sh && yarn makepot",
+    "postdev": "mkdir temp && cp -a dist/* temp && NODE_ENV=production webpack --config webpack.config.js && cp -a temp/* dist && rm -rf temp && .dev/deploy-scripts/tweak-css-files.sh",
     "package": "rm -rf build && mkdir -p build/go && yarn build && rsync -av --exclude-from .distignore --delete . build/go && cd build && zip -r go.zip ./go",
     "version": "grunt version && git add -A .",
     "preversion": "git pull origin master && git reset --hard origin/master",
